issues #8 , #9 : document exported symbols + move dir scanning to another package
