+++ pkg/netpol/eval/internal/k8s/connectionset.go
@@ -21,11 +21,13 @@ import (
 	v1 "k8s.io/api/core/v1"
 )
 
+// ConnectionSet represents a set of allowed connections between two peers on a k8s env
 type ConnectionSet struct {
 	AllowAll         bool
 	AllowedProtocols map[v1.Protocol]*PortSet // map from protocol name to set of allowed ports
 }
 
+// MakeConnectionSet returns a ConnectionSet object with all connections or no connections
 func MakeConnectionSet(all bool) ConnectionSet {
 	if all {
 		return ConnectionSet{AllowAll: true, AllowedProtocols: map[v1.Protocol]*PortSet{}}
@@ -33,6 +35,7 @@ func MakeConnectionSet(all bool) ConnectionSet {
 	return ConnectionSet{AllowedProtocols: map[v1.Protocol]*PortSet{}}
 }
 
+// Intersection updates ConnectionSet object to be the intersection result with other ConnectionSet
 func (conn *ConnectionSet) Intersection(other ConnectionSet) {
 	if other.AllowAll {
 		return
@@ -58,6 +61,7 @@ func (conn *ConnectionSet) Intersection(other ConnectionSet) {
 	}
 }
 
+// IsEmpty returns true if the ConnectionSet has no allowed connections
 func (conn *ConnectionSet) IsEmpty() bool {
 	return !conn.AllowAll && len(conn.AllowedProtocols) == 0
 }
@@ -86,6 +90,7 @@ func (conn *ConnectionSet) checkIfAllConnections() {
 	}
 }
 
+// Union updates ConnectionSet object to be the union result with other ConnectionSet
 func (conn *ConnectionSet) Union(other ConnectionSet) {
 	if conn.AllowAll || other.IsEmpty() {
 		return
@@ -109,6 +114,7 @@ func (conn *ConnectionSet) Union(other ConnectionSet) {
 	conn.checkIfAllConnections()
 }
 
+// Contains returns true if the input port+protocol is an allowed connection
 func (conn *ConnectionSet) Contains(port, protocol string) bool {
 	intPort, err := strconv.Atoi(port)
 	if err != nil {
@@ -122,6 +128,7 @@ func (conn *ConnectionSet) Contains(port, protocol string) bool {
 	return false
 }
 
+// ContainedIn returns true if current ConnectionSet is conatained in the input ConnectionSet object
 func (conn *ConnectionSet) ContainedIn(other ConnectionSet) bool {
 	if other.AllowAll {
 		return true
@@ -141,6 +148,7 @@ func (conn *ConnectionSet) ContainedIn(other ConnectionSet) bool {
 	return true
 }
 
+// AddConnection updates current ConnectionSet object with new allowed connection
 func (conn *ConnectionSet) AddConnection(protocol v1.Protocol, ports PortSet) {
 	if ports.IsEmpty() {
 		return
@@ -153,6 +161,7 @@ func (conn *ConnectionSet) AddConnection(protocol v1.Protocol, ports PortSet) {
 	}
 }
 
+// String returns a string representation of the ConnectionSet object
 func (conn *ConnectionSet) String() string {
 	if conn.AllowAll {
 		return "All Connections"
@@ -167,6 +176,7 @@ func (conn *ConnectionSet) String() string {
 	return strings.Join(resStrings, ",")
 }
 
+// Equal returns true if the current ConnectionSet object is equal to the input object
 func (conn *ConnectionSet) Equal(other ConnectionSet) bool {
 	if conn.AllowAll != other.AllowAll {
 		return false
