+++ pkg/netpol/eval/internal/k8s/ipBlock.go
@@ -18,10 +18,12 @@ const (
 	ipMask   = 0xffffffff
 )
 
+// IPBlock captures a set of ip ranges
 type IPBlock struct {
 	ipRange CanonicalIntervalSet
 }
 
+// ToIPRanges returns a string if the ip ranges in the current IPBlock object
 func (b *IPBlock) ToIPRanges() string {
 	res := ""
 	for index := range b.ipRange.IntervalSet {
@@ -32,6 +34,7 @@ func (b *IPBlock) ToIPRanges() string {
 	return res
 }
 
+// Copy returns a new copy of IPBlock object
 func (b *IPBlock) Copy() *IPBlock {
 	res := &IPBlock{}
 	res.ipRange = b.ipRange.Copy()
@@ -46,7 +49,7 @@ func (b *IPBlock) ipCount() int {
 	return res
 }
 
-// return a set of IpBlock objects, each with a single range of ips
+// split returns a set of IpBlock objects, each with a single range of ips
 func (b *IPBlock) split() []*IPBlock {
 	res := []*IPBlock{}
 	for _, ipr := range b.ipRange.IntervalSet {
@@ -57,7 +60,7 @@ func (b *IPBlock) split() []*IPBlock {
 	return res
 }
 
-//revive:disable:add-constant
+// InttoIP4 returns a string of an ip address from an input integer ip value
 func InttoIP4(ipInt int64) string {
 	// need to do two bit shifting and “0xff” masking
 	b0 := strconv.FormatInt((ipInt>>ipShift0)&ipByte, ipBase)
@@ -67,10 +70,9 @@ func InttoIP4(ipInt int64) string {
 	return b0 + "." + b1 + "." + b2 + "." + b3
 }
 
-//revive:enable:add-constant
-
 // TODO: support allowed_connections considering an ipBlock (cidr), not only ip address
 // TODO: generate tests for disjoint ip blocks (unit tests, and per netpols)
+// DisjointIPBlocks returns an IPBlock of disjoint ip ranges from 2 input IPBlock objects
 func DisjointIPBlocks(set1, set2 []*IPBlock) []*IPBlock {
 	ipbList := []*IPBlock{}
 	for _, ipb := range set1 {
@@ -119,6 +121,7 @@ func addIntervalToList(ipbNew *IPBlock, ipbList []*IPBlock) []*IPBlock {
 	return ipbList
 }
 
+// NewIPBlock returns an IPBlock object from input cidr str an exceptions cidr str
 func NewIPBlock(cidr string, exceptions []string) (*IPBlock, error) {
 	res := IPBlock{ipRange: CanonicalIntervalSet{}}
 	interval, err := cidrToInterval(cidr)
